feat: add lock page flow
